requirements.txt
python-socketio
aiohttp
numpy
python-dotenv